Add env var to avoid pipe deprecation guard
Adds and env var that, if set to true, disables the deprecation guard
for pipe syntax.

When running an ansible playbook, the test [ -p /dev/stdin ] will always
be true. Thus, every 'zplug user/repo' will fail with a warning, even if
the (deprecated) pipe syntax was not used.

Another option would be to log the warning but add the plugin anyway,
i.e. remove the return statement.

Fixes: zplug/zplug#393, zplug/zplug#272